Added City Detail Screen. Implemented: - Weather Api request; - Loading data progress bar - City detail screen design Optimized: - Cities search feature - Added Loading cities progress bar - Async Requests

+++ app/src/main/java/com/slobodyanyuk/testprojectinfotech/data/source/remote/dto/Coordinates.kt
@@ -5,6 +5,6 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class Coordinates(
-    @Json(name = "lon")val longitude: Double?,
-    @Json(name = "lat")val latitude: Double?
+    @Json(name = "lon") val longitude: Float?,
+    @Json(name = "lat") val latitude: Float?
 )
